+++ temba/orgs/integrations/dtone/type.py
@@ -1,6 +1,5 @@
-from django.conf.urls import url
-from django.urls import reverse
-from django.utils.translation import ugettext_lazy as _
+from django.urls import re_path, reverse
+from django.utils.translation import gettext_lazy as _
 
 from ...models import IntegrationType
 from .views import AccountView
@@ -41,4 +40,4 @@ class DTOneType(IntegrationType):
             formax.add_section(self.slug, account_url, icon=self.icon, action="redirect", nobutton=True)
 
     def get_urls(self):
-        return [url(r"^account$", AccountView.as_view(integration_type=self), name="account")]
+        return [re_path(r"^account$", AccountView.as_view(integration_type=self), name="account")]