+++ temba/public/models.py
@@ -1,7 +1,7 @@
 from smartmin.models import SmartModel
 
 from django.db import models
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 
 class Lead(SmartModel):