+++ temba/tickets/models.py
@@ -4,13 +4,13 @@ import regex
 from smartmin.models import SmartModel
 
 from django.conf import settings
-from django.conf.urls import url
 from django.contrib.auth.models import User
 from django.db import models
 from django.db.models import Q, Sum
 from django.template import Engine
+from django.urls import re_path
 from django.utils import timezone
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 from temba import mailroom
 from temba.contacts.models import Contact
@@ -61,7 +61,7 @@ class TicketerType(metaclass=ABCMeta):
         """
         Gets the URL/view configuration for this ticketer's connect page
         """
-        return url(r"^connect", self.connect_view.as_view(ticketer_type=self), name="connect")
+        return re_path(r"^connect", self.connect_view.as_view(ticketer_type=self), name="connect")
 
 
 class Ticketer(SmartModel, DependencyMixin):