+++ temba/utils/analytics/librato/__init__.py
@@ -1,1 +1,0 @@
-from .backend import LibratoBackend  # noqa